Move SRA types from @0xproject/connect to @0xproject/sra-types
