+++ packages/connect/src/http_client.ts
@@ -1,19 +1,10 @@
 import { assert } from '@0xproject/assert';
 import { schemas } from '@0xproject/json-schemas';
-import { SignedOrder } from '@0xproject/types';
-import { fetchAsync } from '@0xproject/utils';
-import * as _ from 'lodash';
-import * as queryString from 'query-string';
-
-import { schemas as clientSchemas } from './schemas/schemas';
 import {
     APIOrder,
     AssetPairsRequestOpts,
     AssetPairsResponse,
-    Client,
     FeeRecipientsResponse,
-    HttpRequestOptions,
-    HttpRequestType,
     OrderbookRequest,
     OrderbookResponse,
     OrderConfigRequest,
@@ -22,7 +13,14 @@ import {
     OrdersResponse,
     PagedRequestOpts,
     RequestOpts,
-} from './types';
+    SignedOrder,
+} from '@0xproject/types';
+import { fetchAsync } from '@0xproject/utils';
+import * as _ from 'lodash';
+import * as queryString from 'query-string';
+
+import { schemas as clientSchemas } from './schemas/schemas';
+import { Client, HttpRequestOptions, HttpRequestType } from './types';
 import { relayerResponseJsonParsers } from './utils/relayer_response_json_parsers';
 
 const TRAILING_SLASHES_REGEX = /\/+$/;
