+++ packages/connect/src/types.ts
@@ -1,5 +1,18 @@
-import { SignedOrder } from '@0xproject/types';
-import { BigNumber } from '@0xproject/utils';
+import {
+    APIOrder,
+    AssetPairsItem,
+    AssetPairsRequestOpts,
+    FeeRecipientsResponse,
+    OrderbookRequest,
+    OrderbookResponse,
+    OrderConfigRequest,
+    OrderConfigResponse,
+    OrdersChannelSubscriptionOpts,
+    OrdersRequestOpts,
+    PagedRequestOpts,
+    PaginatedCollection,
+    SignedOrder,
+} from '@0xproject/types';
 
 export interface Client {
     getAssetPairsAsync: (
@@ -18,138 +31,10 @@ export interface OrdersChannel {
     close: () => void;
 }
 
-/**
- * baseAssetData: The address of assetData designated as the baseToken in the currency pair calculation of price
- * quoteAssetData: The address of assetData designated as the quoteToken in the currency pair calculation of price
- * limit: Maximum number of bids and asks in orderbook snapshot
- */
-export interface OrdersChannelSubscriptionOpts {
-    baseAssetData: string;
-    quoteAssetData: string;
-    limit: number;
-}
-
 export interface OrdersChannelHandler {
     onUpdate: (channel: OrdersChannel, subscriptionOpts: OrdersChannelSubscriptionOpts, orders: APIOrder[]) => void;
     onError: (channel: OrdersChannel, err: Error, subscriptionOpts?: OrdersChannelSubscriptionOpts) => void;
     onClose: (channel: OrdersChannel) => void;
-}
-
-export type OrdersChannelMessage = UpdateOrdersChannelMessage | UnknownOrdersChannelMessage;
-
-export enum OrdersChannelMessageTypes {
-    Update = 'update',
-    Unknown = 'unknown',
-}
-
-export interface UpdateOrdersChannelMessage {
-    type: OrdersChannelMessageTypes.Update;
-    requestId: string;
-    payload: APIOrder[];
-}
-
-export interface UnknownOrdersChannelMessage {
-    type: OrdersChannelMessageTypes.Unknown;
-    requestId: string;
-    payload: undefined;
-}
-
-export enum WebsocketConnectionEventType {
-    Close = 'close',
-    Error = 'error',
-    Message = 'message',
-}
-
-export enum WebsocketClientEventType {
-    Connect = 'connect',
-    ConnectFailed = 'connectFailed',
-}
-
-export type OrdersResponse = PaginatedCollection<APIOrder>;
-
-export interface APIOrder {
-    order: SignedOrder;
-    metaData: object;
-}
-
-export interface AssetPairsRequestOpts {
-    assetDataA?: string;
-    assetDataB?: string;
-}
-
-export type AssetPairsResponse = PaginatedCollection<AssetPairsItem>;
-
-export interface AssetPairsItem {
-    assetDataA: Asset;
-    assetDataB: Asset;
-}
-
-export interface Asset {
-    assetData: string;
-    minAmount: BigNumber;
-    maxAmount: BigNumber;
-    precision: number;
-}
-
-export interface OrdersRequestOpts {
-    makerAssetProxyId?: string;
-    takerAssetProxyId?: string;
-    makerAssetAddress?: string;
-    takerAssetAddress?: string;
-    exchangeAddress?: string;
-    senderAddress?: string;
-    makerAssetData?: string;
-    takerAssetData?: string;
-    makerAddress?: string;
-    takerAddress?: string;
-    traderAddress?: string;
-    feeRecipientAddress?: string;
-}
-
-export interface OrderbookRequest {
-    baseAssetData: string;
-    quoteAssetData: string;
-}
-
-export interface OrderbookResponse {
-    bids: PaginatedCollection<APIOrder>;
-    asks: PaginatedCollection<APIOrder>;
-}
-
-export interface PaginatedCollection<T> {
-    total: number;
-    page: number;
-    perPage: number;
-    records: T[];
-}
-
-export interface OrderConfigRequest {
-    makerAddress: string;
-    takerAddress: string;
-    makerAssetAmount: BigNumber;
-    takerAssetAmount: BigNumber;
-    makerAssetData: string;
-    takerAssetData: string;
-    exchangeAddress: string;
-    expirationTimeSeconds: BigNumber;
-}
-
-export interface OrderConfigResponse {
-    makerFee: BigNumber;
-    takerFee: BigNumber;
-    feeRecipientAddress: string;
-    senderAddress: string;
-}
-
-export type FeeRecipientsResponse = PaginatedCollection<string>;
-
-export interface RequestOpts {
-    networkId?: number;
-}
-
-export interface PagedRequestOpts {
-    page?: number;
-    perPage?: number;
 }
 
 export interface HttpRequestOptions {
