+++ packages/connect/test/fixtures/standard_relayer_api/fee_recipients.ts
@@ -1,4 +1,4 @@
-import { FeeRecipientsResponse } from '../../../src/types';
+import { FeeRecipientsResponse } from '@0xproject/types';
 
 export const feeRecipientsResponse: FeeRecipientsResponse = {
     total: 3,
